requirements.txt
pygame
pyYAML
customtkinter
